+++ interface/lib/plugins/ldap_auth_plugin/sql/mail_user.sql
@@ -1,2 +1,2 @@
 ALTER TABLE mail_user
-ADD COLUMN IF NOT EXISTS ldap_enabled ENUM('n','y') NOT NULL DEFAULT 'n';
+ADD COLUMN IF NOT EXISTS ldap_enabled ENUM('n','y') NOT NULL DEFAULT 'y';